refactor: Restructure application by moving Job and Item definitions to separate modules
