+++ src/main.rs
@@ -1,20 +1,22 @@
 mod application;
 mod constants;
 mod file_util;
+mod item;
+mod job;
 mod sysexits;
 
-use crate::application::{Application, BackupModel, CompressFormat, Level};
+use crate::application::{Application, init_config};
 use crate::file_util::*;
-use anyhow::{Context, Result, anyhow};
-use application::{Job, init_config};
+use crate::item::{get_item, get_items};
+use crate::job::{BackupModel, CompressFormat, Job, Level};
+use anyhow::{Result, anyhow};
 use clap::{Parser, Subcommand, ValueEnum};
 use futures::StreamExt;
 use futures::stream::FuturesUnordered;
 use std::io::{ErrorKind, Write};
-use std::path::{Path, PathBuf};
+use std::path::PathBuf;
 use std::{fs, io, process};
 use tokio::runtime::Builder;
-use walkdir::WalkDir;
 
 /// Entry point for the hbackup CLI application.
 /// Parses command-line arguments and dispatches to the appropriate command handler.
@@ -679,148 +681,6 @@ fn rollback_config_file() {
     }
 }
 
-#[derive(Debug)]
-pub(crate) struct Item {
-    pub src: PathBuf,
-    pub dest: PathBuf,
-    pub strategy: Strategy,
-}
-
-impl Item {
-    fn new(src: PathBuf, dest: PathBuf, strategy: Strategy) -> Item {
-        Item {
-            src,
-            dest,
-            strategy,
-        }
-    }
-
-    fn from_delete_strategy(dest: PathBuf) -> Item {
-        Self::new(PathBuf::new(), dest, Strategy::Delete)
-    }
-
-    fn from_copy_strategy(src: PathBuf, dest: PathBuf) -> Item {
-        Self::new(src, dest, Strategy::Copy)
-    }
-
-    fn from_ignore_strategy(src: PathBuf, dest: PathBuf) -> Item {
-        Self::new(src, dest, Strategy::Ignore)
-    }
-
-    fn from_notupdate_strategy(src: PathBuf, dest: PathBuf) -> Item {
-        Self::new(src, dest, Strategy::NotUpdate)
-    }
-
-    fn change_delete_strategy(&mut self) {
-        self.src = PathBuf::new();
-        self.strategy = Strategy::Delete;
-    }
-}
-
-#[derive(PartialEq, Debug)]
-pub(crate) enum Strategy {
-    Copy,
-    Ignore,
-    NotUpdate,
-    Delete,
-}
-
-fn get_item(job: Job) -> Result<Item> {
-    let src = job.source;
-    if !src.exists() {
-        eprintln!("The path {src:?} is not exists");
-        process::exit(1);
-    } else if !src.is_file() {
-        eprintln!("The path {src:?} is not file");
-        process::exit(1);
-    }
-
-    let dest = job.target;
-    let dest = if dest.exists() && dest.is_dir() {
-        let file_name = src.file_name().with_context(|| "Invalid file name")?;
-        dest.join(file_name)
-    } else {
-        dest
-    };
-    let model = job.model.unwrap_or_default();
-    match model {
-        BackupModel::Full => Ok(Item::new(src, dest, Strategy::Copy)),
-        BackupModel::Mirror => {
-            if needs_update(&src, &dest)? {
-                Ok(Item::new(src, dest, Strategy::Copy))
-            } else {
-                Ok(Item::new(src, dest, Strategy::NotUpdate))
-            }
-        }
-    }
-}
-
-fn get_items(job: Job) -> Result<Vec<Item>> {
-    let src = job.source;
-    if !src.exists() {
-        eprintln!("The path {src:?} is not exists");
-        process::exit(1);
-    } else if !src.is_dir() {
-        eprintln!("The path {src:?} is not directory");
-        process::exit(1);
-    }
-
-    let target = job.target;
-    fs::create_dir_all(&target)?;
-
-    let model = job.model.unwrap_or_default();
-    let prefix = src.parent().unwrap_or_else(|| Path::new(""));
-    let mut vec = vec![];
-    let ignore_paths: Vec<_> = job
-        .ignore
-        .as_ref()
-        .map(|dirs| dirs.iter().map(|s| src.join(s)).collect())
-        .unwrap_or_default();
-
-    for entry in WalkDir::new(&src) {
-        let entry = entry?;
-        let entry_path = entry.path();
-        if entry_path.is_file() {
-            let rel = entry_path.strip_prefix(prefix)?;
-            let dest = target.join(rel);
-            if ignore_paths.iter().any(|p| entry_path.starts_with(p)) {
-                vec.push(Item::from_ignore_strategy(entry_path.to_path_buf(), dest));
-                continue;
-            }
-            match model {
-                BackupModel::Full => {
-                    vec.push(Item::from_copy_strategy(entry_path.to_path_buf(), dest));
-                }
-                BackupModel::Mirror => {
-                    if needs_update(entry_path, &dest)? {
-                        vec.push(Item::from_copy_strategy(entry_path.to_path_buf(), dest));
-                    } else {
-                        vec.push(Item::from_notupdate_strategy(
-                            entry_path.to_path_buf(),
-                            dest,
-                        ));
-                    }
-                }
-            }
-        }
-    }
-    for entry in WalkDir::new(&target) {
-        let entry = entry?;
-        let entry_path = entry.path();
-        if entry_path.is_file() {
-            if let Some(i) = vec.iter().position(|v| v.dest.eq(entry_path)) {
-                if vec[i].strategy == Strategy::Ignore {
-                    vec[i].change_delete_strategy();
-                }
-            } else {
-                vec.push(Item::from_delete_strategy(entry_path.to_path_buf()));
-            }
-        }
-    }
-
-    Ok(vec)
-}
-
 /// Returns the canonical, absolute form of the path with all intermediate
 /// components normalized and symbolic links resolved.
 fn canonicalize(path: PathBuf) -> PathBuf {
@@ -845,8 +705,8 @@ fn canonicalize(path: PathBuf) -> PathBuf {
 
 #[cfg(test)]
 mod tests {
-    use crate::application::{CompressFormat, Job, Level};
     use crate::display_jobs;
+    use crate::job::{CompressFormat, Job, Level};
     use std::path::PathBuf;
 
     #[test]
